Changes for Move exec operations from ContainerOperations
